feat(mn): mnDiagram_8024227C 95.39 -> 96.03 (permuter: u64 cursor temp)
Permuter triage round 3 (coder1 job mnDiagram_8024227C-coder1-20260612-013111,
~143k iters, best score 1345 vs base 1560). The winning lever widens the local
cursor temp var_r0_2 from s32 to unsigned long long. var_r0_2 only ever holds a
(u8)-masked cursor value (0..255), always re-cast (u8) or passed by value at its
two uses (mnDiagram_GetNameTotalKOs and GetPersistentNameData), so the wider
declaration is value-identical (no upper-bit garbage is ever read) -- it only
changes register coloring on the indexed-struct-pointer-materialization path.

Verified: report.json fuzzy 95.389 -> 96.033, sole changed function; all 53
mnDiagram 100%s hold (incl. mnDiagram_CursorProc); no regressions across the TUs.

+++ src/melee/mn/mndiagram.c
@@ -2430,7 +2430,7 @@ void mnDiagram_8024227C(void* arg0, s32 arg1, s32 arg2, u8 arg3)
     s32 var_r3;
     s32 var_r17_4;
     s32 var_r19_5;
-    s32 var_r0_2;
+    unsigned long long var_r0_2;
     s32 var_r17_6;
     s32 var_r23;
     u8 var_r24;
